pyproject.toml
[project]
name = "enzax"
version = "0.1.1"
description = "Differentiable models of enzyme-catalysed reaction networks"
authors = [
    {name = "Teddy Groves", email = "tedgro@dtu.dk"},
]
dependencies = [
    "blackjax>=1.2.1",
    "diffrax>=0.6.0",
    "jaxtyping>=0.2.31",
    "arviz>=0.19.0",
]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[project.optional-dependencies]
docs = [
    "mkdocs>=1.6.0",
    "mkdocs-material>=9.5.32",
    "mkdocstrings>=0.25.2",
    "mkdocstrings-python>=1.10.8",
    "pymdown-extensions>=10.9",
]
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pdm]
distribution = true

[tool.pdm.dev-dependencies]
dev = [
    "pytest>=8.3.2",
    "pytest-cov>=5.0.0",
    "pre-commit>=3.8.0",
]
[tool.ruff]
line-length = 80

[tool.ruff.lint]
ignore = ["F722"]
extend-select = ["E501"]  # line length is checked

[tool.ruff.lint.isort]
known-first-party = ["enzax"]
